rename main to main_cli
