+++ functions/scraping-events/main_dev.py
@@ -4,6 +4,6 @@ if __name__ == "__main__":
     # load env vars before even importing anything from `scraping_events`
     dotenv.load_dotenv()
 
-    from scraping_events.main import main
+    from scraping_events.main_cli import main
 
     main()
